# ui/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "quarry-server",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.7-nightly.20260731"
  },
  "paths": {
...
    "/v1/libraries/{library}/documents/{path}/review": {
      "get": {
        "tags": [
          "review"
        ],
        "operationId": "document_review_openapi",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeResolved",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DryRunValue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentReviewResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tmp/documents/{secret}/review": {
      "get": {
        "tags": [
          "review"
        ],
        "operationId": "tmp_document_review_openapi",
        "parameters": [
          {
            "name": "secret",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeResolved",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DryRunValue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentReviewResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentBlockRef": {
        "type": "object",
        "properties": {
          "blockId": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "AgentReviewComment": {
        "type": "object",
        "required": [
          "id",
          "status",
          "by",
          "at",
          "ref",
          "quote",
          "body",
          "replies"
        ],
        "properties": {
          "anchor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BlockReviewAnchor",
                "description": "Row-anchored position; present only when the document has canonical\nblock rows (the Phase 2 review projection)."
              }
            ]
          },
          "at": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "by": {
            "type": "string"
          },
          "editedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "quote": {
            "type": "string"
          },
          "ref": {
            "$ref": "#/components/schemas/AgentBlockRef"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentReviewReply"
            }
          },
          "status": {
            "type": "string"
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResolvedTarget",
                "description": "Native targets can span blocks. Discussion status is independent of\nwhether target text is attached, hidden, deleted, or unknown at import."
              }
            ]
          }
        }
      },
      "AgentReviewConflict": {
        "type": "object",
        "description": "A `kind = conflict` review item: a diff3 merge kept the canonical side and\nrecorded the losing incoming hunk here. `conflict.keep_canonical` resolves\nwithout a content change; `conflict.accept_incoming` verifies and replaces\nthe retained hunk atomically.",
        "required": [
          "id",
          "status",
          "by",
          "at",
          "baseMarkdown",
          "incomingMarkdown",
          "canonicalMarkdown"
        ],
        "properties": {
          "afterBlockId": {
            "type": [
              "string",
              "null"
            ],
            "description": "The surviving block the conflict region attaches after; `null` means\nthe document start."
          },
          "at": {
            "type": "string"
          },
          "baseMarkdown": {
            "type": "string",
            "description": "The base (shadow) context the merge diffed against."
          },
          "by": {
            "type": "string"
          },
          "canonicalMarkdown": {
            "type": "string",
            "description": "The canonical side that was retained (empty = canonical had deleted\nthe region)."
          },
          "id": {
            "type": "string"
          },
          "incomingMarkdown": {
            "type": "string",
            "description": "The losing incoming hunk (empty = the write deleted this region)."
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentReviewReply"
            }
          },
          "status": {
            "type": "string"
          }
        }
      },
      "AgentReviewReply": {
        "type": "object",
        "required": [
          "id",
          "status",
          "by",
          "at",
          "body"
        ],
        "properties": {
          "at": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "by": {
            "type": "string"
          },
          "editedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResolvedTarget"
              }
            ]
          }
        }
      },
      "AgentReviewResponse": {
        "type": "object",
        "required": [
          "documentId",
          "baseToken",
          "comments",
          "suggestions",
          "conflicts"
        ],
        "properties": {
          "baseToken": {
            "type": "string"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentReviewComment"
            }
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentReviewConflict"
            },
            "description": "diff3 conflict review items (Phase 4): unresolved whole-file merge\nconflicts, present only for documents with canonical block rows."
          },
          "documentId": {
            "type": "string"
          },
          "suggestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentReviewSuggestion"
            }
          }
        }
      },
      "AgentReviewSuggestion": {
        "type": "object",
        "required": [
          "id",
          "status",
          "kind",
          "by",
          "at",
          "ref",
          "quote",
          "content",
          "preview",
          "replies"
        ],
        "properties": {
          "acceptance_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "action": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProposalAction"
              }
            ]
          },
          "anchor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BlockReviewAnchor",
                "description": "Row-anchored position; present only when the document has canonical\nblock rows (the Phase 2 review projection)."
              }
            ]
          },
          "at": {
            "type": "string"
          },
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "by": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AgentSuggestionKind"
          },
          "preview": {
            "$ref": "#/components/schemas/AgentSuggestionPreview"
          },
          "quote": {
            "type": "string"
          },
          "ref": {
            "$ref": "#/components/schemas/AgentBlockRef"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentReviewReply"
            }
          },
          "status": {
            "type": "string"
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResolvedTarget"
              }
            ]
          }
        }
      },
...
      "AgentSuggestionKind": {
        "type": "string",
        "enum": [
          "block_move",
          "block_update",
          "format",
          "block_delete",
          "markdown_insert",
          "insert",
          "delete",
          "remove",
          "replace",
          "substitution"
        ]
      },
      "AgentSuggestionPreview": {
        "type": "object",
        "required": [
          "before",
          "after"
        ],
        "properties": {
          "after": {
            "type": "string"
          },
          "before": {
            "type": "string"
          }
        }
      },
      "ApiErrorCode": {
        "type": "string",
        "description": "Stable machine-readable codes returned by every `/v1` HTTP error.",
        "enum": [
          "FORBIDDEN",
          "INVALID_REQUEST",
          "NOT_FOUND",
          "GONE",
          "PRECONDITION_REQUIRED",
          "PRECONDITION_FAILED",
          "CONFLICT",
          "METHOD_NOT_ALLOWED",
          "UNSUPPORTED_MEDIA_TYPE",
          "PAYLOAD_TOO_LARGE",
          "UNPROCESSABLE_ENTITY",
          "SERVICE_BUSY",
          "INTERNAL_ERROR",
          "STALE_BASE",
          "BLOCK_DELETED",
          "ANCHOR_NOT_FOUND",
          "BLOCK_MOVE_CONFLICT",
          "SUGGESTION_INVALIDATED",
          "SUGGESTION_ALREADY_RESOLVED",
          "UNSUPPORTED_MARKDOWN",
          "INVALID_TRANSACTION",
          "UNKNOWN_BLOCK_TYPE",
          "UNSUPPORTED_BLOCK_DOCUMENT"
        ]
      },
      "ApiErrorDetails": {
        "type": "object",
        "description": "Machine-actionable context for a failed request. Transaction failures\nidentify the exact op and target; validation failures may additionally\nidentify the field, rejected value, current value, or allowed values.",
        "properties": {
          "allowed_values": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Accepted values for a closed vocabulary."
          },
          "current_value": {
            "type": [
              "string",
              "null"
            ],
            "description": "Current server value, when useful for rebuilding a request."
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "Request field that failed validation or a precondition."
          },
          "op": {
            "type": [
              "string",
              "null"
            ],
            "description": "Operation discriminator, such as `replace_block_content`."
          },
          "op_index": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Zero-based index into a transaction's `ops` array.",
            "minimum": 0
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiErrorTarget",
                "description": "Addressable entity involved in the failure."
              }
            ]
          },
          "value": {
            "type": [
              "string",
              "null"
            ],
            "description": "Rejected request value."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "retryable",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiErrorDetails"
              }
            ]
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ApiErrorTarget": {
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable identifier copied from the request or current review state."
          },
          "kind": {
            "type": "string",
            "description": "Entity category, such as `block`, `review_item`, or `conflict`."
          }
        }
      },
...
      "BlockReviewAnchor": {
        "type": "object",
        "description": "Row-anchored review position, attached to review items projected from\nblock rows. Offsets are UTF-16 code units; `end_offset` is exclusive.",
        "required": [
          "blockId",
          "startOffset",
          "endOffset"
        ],
        "properties": {
          "blockId": {
            "type": "string"
          },
          "endOffset": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "startOffset": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "DryRunValue": {
        "type": "string",
        "enum": [
          "1",
          "true",
          "yes",
          "0",
          "false",
          "no"
        ]
      },
...
      "ProposalAction": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "reason",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "unavailable"
                ]
              },
              "reason": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "at",
              "delete_target",
              "original_quote",
              "kind"
            ],
            "properties": {
              "at": {
                "$ref": "#/components/schemas/TextPoint"
              },
              "delete_target": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TargetFragment"
                }
              },
              "kind": {
                "type": "string",
                "enum": [
                  "text"
                ]
              },
              "original_quote": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "target",
              "name",
              "value",
              "expected",
              "kind"
            ],
            "properties": {
              "expected": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": false,
                  "prefixItems": [
                    {
                      "type": "string"
                    },
                    {}
                  ]
                },
                "description": "Text and this mark's values at proposal time. Other marks and\nblock placement can change without invalidating the decision."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "format"
                ]
              },
              "name": {
                "type": "string"
              },
              "target": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TargetFragment"
                }
              },
              "value": {}
            }
          },
          {
            "type": "object",
            "required": [
              "block",
              "block_kind",
              "attrs",
              "expected_kind",
              "expected_attrs",
              "kind"
            ],
            "properties": {
              "attrs": {
                "type": "object",
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              },
              "block": {
                "type": "string"
              },
              "block_kind": {
                "type": "string"
              },
              "expected_attrs": {
                "type": "object",
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              },
              "expected_kind": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "update_block"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "block",
              "kind"
            ],
            "properties": {
              "before": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "block": {
                "type": "string"
              },
              "expected_before": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "expected_parent": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "kind": {
                "type": "string",
                "enum": [
                  "move_block"
                ]
              },
              "parent": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "block",
              "expected",
              "kind"
            ],
            "properties": {
              "block": {
                "type": "string"
              },
              "expected": {
                "type": "string",
                "description": "Exact subtree content when the deletion was proposed. A changed\nsubtree requires a new decision instead of dropping unseen edits."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "delete_block"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "blocks",
              "kind"
            ],
            "properties": {
              "before": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "blocks": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Block"
                }
              },
              "kind": {
                "type": "string",
                "enum": [
                  "insert_blocks"
                ]
              },
              "parent": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        ]
      },
...
      "ResolvedTarget": {
        "type": "object",
        "required": [
          "state",
          "attachments"
        ],
        "properties": {
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "state": {
            "$ref": "#/components/schemas/TargetState"
          }
        }
      },
...
      "TargetState": {
        "type": "string",
        "enum": [
          "attached",
          "hidden",
          "deleted",
          "unattached"
        ]
      },
...
    }
  }
}